# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/stablestack/banks": {
      "get": {
        "operationId": "StablestackController_getBanks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of banks"
          }
        },
        "summary": "Get list of banks",
        "tags": [
          "Stablestack"
        ],
        "security": [
          {
            "API-Key": []
          }
        ]
      }
    },
    "/stablestack/resolve-account": {
      "get": {
        "description": "Resolves account name from bank code and account number. Requires authentication (JWT token in Authorization header or API key in x-api-key header).",
        "operationId": "StablestackController_resolveAccount",
        "parameters": [
          {
            "name": "bankCode",
            "required": true,
            "in": "query",
            "description": "Bank code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountNumber",
            "required": true,
            "in": "query",
            "description": "Account number",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account name resolved successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "API-Key": []
          }
        ],
        "summary": "Resolve account name",
        "tags": [
          "Stablestack"
        ]
      }
    },
    "/stablestack/onramp": {
      "post": {
        "description": "Requires authentication (JWT token in Authorization header or API key in x-api-key header)",
        "operationId": "StablestackController_onRamp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/onRampDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Onramp Initialisation response"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "API-Key": []
          }
        ],
        "summary": "Initialise onramp transaction",
        "tags": [
          "Stablestack"
        ]
      }
    },
    "/stablestack/offramp": {
      "post": {
        "description": "Requires authentication (JWT token in Authorization header or API key in x-api-key header)",
        "operationId": "StablestackController_offRamp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/offRampDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Offramp Initialisation response"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "API-Key": []
          }
        ],
        "summary": "Initialise offramp transaction",
        "tags": [
          "Stablestack"
        ]
      }
    },
    "/stablestack/offramp/{reference}/confirm-deposit": {
      "post": {
        "description": "Reports the Stellar transaction hash of a CNGN deposit sent to the offramp collection account. Verifies it on-chain via Horizon before moving the transaction out of PENDING.",
        "operationId": "StablestackController_confirmOfframpDeposit",
        "parameters": [
          {
            "name": "reference",
            "required": true,
            "in": "path",
            "description": "Offramp transaction reference",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transactionHash": {
                    "type": "string"
                  }
                },
                "required": [
                  "transactionHash"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deposit confirmed, transaction moved to PROCESSING"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "API-Key": []
          }
        ],
        "summary": "Confirm a CNGN deposit for an offramp transaction",
        "tags": [
          "Stablestack"
        ]
      }
    },
    "/stablestack/transactions": {
      "get": {
        "description": "Requires authentication (JWT token in Authorization header or API key in x-api-key header) and verified KYC status. Returns transactions for the authenticated user.",
        "operationId": "StablestackController_getTransactions",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reference",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 10)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of transactions"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "API-Key": []
          }
        ],
        "summary": "Get ramp transactions",
        "tags": [
          "Stablestack"
        ]
      }
    },
    "/stablestack/webhook": {
      "post": {
        "operationId": "WebhookController_handleWebhook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully"
          },
          "404": {
            "description": "Transaction not found"
          }
        },
        "summary": "Receive webhook from Flint API",
        "tags": [
          "Stablestack"
        ],
        "security": [
          {
            "API-Key": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "CNGN Ramp Business API",
    "description": "Programmatic API for businesses to perform onramp, offramp, and swap operations.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Your business API key generated from the admin portal."
      }
    },
    "schemas": {
...
      "OnRampDestinationDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "address"
        ]
      },
      "onRampDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "on",
            "default": "on"
          },
          "network": {
            "type": "string",
            "example": "stellar"
          },
          "amount": {
            "type": "number",
            "example": 1
          },
          "destination": {
            "$ref": "#/components/schemas/OnRampDestinationDto"
          },
          "notifyUrl": {
            "type": "string",
            "example": ""
          }
        },
        "required": [
          "type",
          "network",
          "amount",
          "destination"
        ]
      },
      "offRampDestinationDto": {
        "type": "object",
        "properties": {
          "bankCode": {
            "type": "string",
            "example": "string"
          },
          "accountNumber": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "bankCode",
          "accountNumber"
        ]
      },
      "offRampDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "off",
            "default": "off"
          },
          "network": {
            "type": "string",
            "example": "stellar"
          },
          "amount": {
            "type": "number",
            "example": 1
          },
          "destination": {
            "$ref": "#/components/schemas/offRampDestinationDto"
          },
          "notifyUrl": {
            "type": "string",
            "example": ""
          }
        },
        "required": [
          "type",
          "network",
          "amount",
          "destination"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      }
    }
  }
}